Refactor and separate functions
