+++ resources/lib/__init__.py
@@ -1,0 +1,1 @@
+# Dummy file to make this directory a package.
